web/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
httpx==0.28.*
websockets==15.*
python-multipart==0.0.*
facenet-pytorch==2.6.*
ultralytics>=8.0
torch
torchvision
opencv-python
numpy
pillow
